feat(app): move signals
